+++ src/main/java/com/example/joyerialu/ObjetoAlmacen.java
@@ -4,11 +4,11 @@ public class ObjetoAlmacen {
 
     private int id;
     private String nombre;
-    private int precio;
+    private double precio;
     private String descripcion;
     private int stock;
 
-    public ObjetoAlmacen(int id, String nombre, int precio, String descripcion, int stock) {
+    public ObjetoAlmacen(int id, String nombre, double precio, String descripcion, int stock) {
         this.id = id;
         this.nombre = nombre;
         this.precio = precio;
@@ -24,7 +24,7 @@ public class ObjetoAlmacen {
         return nombre;
     }
 
-    public int getPrecio() {
+    public double getPrecio() {
         return precio;
     }
